Blog: Add Featured Section (#11)
Add a new section to the blog page, which features the two most recent
blog posts. For these two blog posts we are showing the image, title and
the description. All other blog posts are then displayed in a list below
these featured blog posts as it was already done.

This commit also fixes a small layout bug in the about page where the
div arround the projects table caused an additional padding, which was
unintended. This is now removed by setting the top and bottom margin of
the table to zero.