Generate PNGs for Cheat Sheets Using Puppeteer
Similar to the creation of PDFs, we are now also creating the PNGs of
the cheat sheets using Puppeteer. The PNGs can not be completly removed,
because we need them for the preview images of an cheat sheet. A user
can now also decide if he wants to download the PDF or PNG version of a
cheat sheet.